+++ cdap-app-fabric/src/main/java/co/cask/cdap/adapter/Sink.java
@@ -86,11 +86,7 @@ public final class Sink {
 
     Sink sink = (Sink) o;
 
-    if (name.equals(sink.name) && properties.equals(sink.properties) &&  type == sink.type) {
-      return true;
-    } else {
-      return false;
-    }
+    return (name.equals(sink.name) && properties.equals(sink.properties) &&  type == sink.type);
   }
 
   @Override
